core/requirements.txt
django
django-ninja
uvicorn
whitenoise
python-dotenv
django-cors-headers
python-jose
google-cloud-pubsub
requests
google-auth
